fix(cleanup): newlines and whitespace
