+++ lib/base-command.js
@@ -179,4 +179,5 @@ class BaseCommand {
     this.workspacePaths = [...ws.values()]
   }
 }
+
 module.exports = BaseCommand
