+++ lib/commands/adduser.js
@@ -1,7 +1,6 @@
 const { log, output } = require('proc-log')
 const { redactLog: replaceInfo } = require('@npmcli/redact')
 const auth = require('../utils/auth.js')
-
 const BaseCommand = require('../base-command.js')
 
 class AddUser extends BaseCommand {
@@ -47,4 +46,5 @@ class AddUser extends BaseCommand {
     output.standard(message)
   }
 }
+
 module.exports = AddUser
