+++ lib/commands/ci.js
@@ -3,7 +3,6 @@ const runScript = require('@npmcli/run-script')
 const fs = require('fs/promises')
 const { log, time } = require('proc-log')
 const validateLockfile = require('../utils/validate-lockfile.js')
-
 const ArboristWorkspaceCmd = require('../arborist-cmd.js')
 
 class CI extends ArboristWorkspaceCmd {
