+++ lib/commands/config.js
@@ -6,6 +6,7 @@ const localeCompare = require('@isaacs/string-locale-compare')('en')
 const pkgJson = require('@npmcli/package-json')
 const { defaults, definitions } = require('@npmcli/config/lib/definitions')
 const { log, output } = require('proc-log')
+const BaseCommand = require('../base-command.js')
 
 // These are the configs that we can nerf-dart. Not all of them currently even
 // *have* config definitions so we have to explicitly validate them here
@@ -46,7 +47,6 @@ const publicVar = k => {
   return true
 }
 
-const BaseCommand = require('../base-command.js')
 class Config extends BaseCommand {
   static description = 'Manage the npm configuration files'
   static name = 'config'
