+++ lib/commands/dist-tag.js
@@ -205,4 +205,5 @@ class DistTag extends BaseCommand {
     return data
   }
 }
+
 module.exports = DistTag
