+++ lib/commands/doctor.js
@@ -8,6 +8,7 @@ const semver = require('semver')
 const { log, output } = require('proc-log')
 const ping = require('../utils/ping.js')
 const { defaults } = require('@npmcli/config/lib/definitions')
+const BaseCommand = require('../base-command.js')
 
 const maskLabel = mask => {
   const label = []
@@ -93,7 +94,7 @@ const subcommands = [
   //   - verify all local packages have bins linked
   // What is the fix for these?
 ]
-const BaseCommand = require('../base-command.js')
+
 class Doctor extends BaseCommand {
   static description = 'Check the health of your npm environment'
   static name = 'doctor'
