+++ lib/commands/edit.js
@@ -1,34 +1,31 @@
-// npm edit <pkg>
-// open the package folder in the $EDITOR
-
 const { resolve } = require('path')
 const { lstat } = require('fs/promises')
 const cp = require('child_process')
 const completion = require('../utils/installed-shallow.js')
 const BaseCommand = require('../base-command.js')
 
-const splitPackageNames = (path) => {
-  return path.split('/')
-    // combine scoped parts
-    .reduce((parts, part) => {
-      if (parts.length === 0) {
-        return [part]
-      }
+const splitPackageNames = (path) => path.split('/')
+// combine scoped parts
+  .reduce((parts, part) => {
+    if (parts.length === 0) {
+      return [part]
+    }
 
-      const lastPart = parts[parts.length - 1]
-      // check if previous part is the first part of a scoped package
-      if (lastPart[0] === '@' && !lastPart.includes('/')) {
-        parts[parts.length - 1] += '/' + part
-      } else {
-        parts.push(part)
-      }
+    const lastPart = parts[parts.length - 1]
+    // check if previous part is the first part of a scoped package
+    if (lastPart[0] === '@' && !lastPart.includes('/')) {
+      parts[parts.length - 1] += '/' + part
+    } else {
+      parts.push(part)
+    }
 
-      return parts
-    }, [])
-    .join('/node_modules/')
-    .replace(/(\/node_modules)+/, '/node_modules')
-}
+    return parts
+  }, [])
+  .join('/node_modules/')
+  .replace(/(\/node_modules)+/, '/node_modules')
 
+// npm edit <pkg>
+// open the package folder in the $EDITOR
 class Edit extends BaseCommand {
   static description = 'Edit an installed package'
   static name = 'edit'
@@ -63,4 +60,5 @@ class Edit extends BaseCommand {
     })
   }
 }
+
 module.exports = Edit
