+++ lib/commands/explain.js
@@ -126,4 +126,5 @@ class Explain extends ArboristWorkspaceCmd {
     })
   }
 }
+
 module.exports = Explain
