+++ lib/commands/explore.js
@@ -1,6 +1,3 @@
-// npm explore <pkg>[@<version>]
-// open a subshell to the package folder.
-
 const pkgJson = require('@npmcli/package-json')
 const runScript = require('@npmcli/run-script')
 const { join, relative } = require('path')
@@ -8,6 +5,8 @@ const { log, output } = require('proc-log')
 const completion = require('../utils/installed-shallow.js')
 const BaseCommand = require('../base-command.js')
 
+// npm explore <pkg>[@<version>]
+// open a subshell to the package folder.
 class Explore extends BaseCommand {
   static description = 'Browse an installed package'
   static name = 'explore'
@@ -71,4 +70,5 @@ class Explore extends BaseCommand {
     })
   }
 }
+
 module.exports = Explore
