+++ lib/commands/fund.js
@@ -222,4 +222,5 @@ class Fund extends ArboristWorkspaceCmd {
     return [url, message]
   }
 }
+
 module.exports = Fund
