+++ lib/commands/get.js
@@ -19,4 +19,5 @@ class Get extends BaseCommand {
     return this.npm.exec('config', ['get'].concat(args))
   }
 }
+
 module.exports = Get
