+++ lib/commands/help-search.js
@@ -191,4 +191,5 @@ class HelpSearch extends BaseCommand {
     return finalOut.trim()
   }
 }
+
 module.exports = HelpSearch
