+++ lib/commands/help.js
@@ -5,9 +5,9 @@ const { glob } = require('glob')
 const { output } = require('proc-log')
 const localeCompare = require('@isaacs/string-locale-compare')('en')
 const { deref } = require('../utils/cmd-list.js')
+const BaseCommand = require('../base-command.js')
 
 const globify = pattern => pattern.split('\\').join('/')
-const BaseCommand = require('../base-command.js')
 
 // Strips out the number from foo.7 or foo.7. or foo.7.tgz
 // We don't currently compress our man pages but if we ever did this would
@@ -111,4 +111,5 @@ class Help extends BaseCommand {
     return 'file:///' + path.resolve(this.npm.npmRoot, `docs/output/${sect}/${f}.html`)
   }
 }
+
 module.exports = Help
