+++ lib/commands/hook.js
@@ -2,8 +2,8 @@ const hookApi = require('libnpmhook')
 const otplease = require('../utils/otplease.js')
 const relativeDate = require('tiny-relative-date')
 const { output } = require('proc-log')
-
 const BaseCommand = require('../base-command.js')
+
 class Hook extends BaseCommand {
   static description = 'Manage registry hooks'
   static name = 'hook'
@@ -105,4 +105,5 @@ class Hook extends BaseCommand {
     return `${hook.type === 'owner' ? '~' : ''}${hook.name}`
   }
 }
+
 module.exports = Hook
