+++ lib/commands/install-ci-test.js
@@ -1,8 +1,7 @@
-// npm install-ci-test
-// Runs `npm ci` and then runs `npm test`
-
 const CI = require('./ci.js')
 
+// npm install-ci-test
+// Runs `npm ci` and then runs `npm test`
 class InstallCITest extends CI {
   static description = 'Install a project with a clean slate and run tests'
   static name = 'install-ci-test'
@@ -12,4 +11,5 @@ class InstallCITest extends CI {
     return this.npm.exec('test', [])
   }
 }
+
 module.exports = InstallCITest
