+++ lib/commands/install-test.js
@@ -1,8 +1,7 @@
-// npm install-test
-// Runs `npm install` and then runs `npm test`
-
 const Install = require('./install.js')
 
+// npm install-test
+// Runs `npm install` and then runs `npm test`
 class InstallTest extends Install {
   static description = 'Install package(s) and run tests'
   static name = 'install-test'
@@ -12,4 +11,5 @@ class InstallTest extends Install {
     return this.npm.exec('test', [])
   }
 }
+
 module.exports = InstallTest
