+++ lib/commands/link.js
@@ -185,4 +185,5 @@ class Link extends ArboristWorkspaceCmd {
     return missing
   }
 }
+
 module.exports = Link
