+++ lib/commands/logout.js
@@ -46,4 +46,5 @@ class Logout extends BaseCommand {
     await this.npm.config.save(level)
   }
 }
+
 module.exports = Logout
