+++ lib/commands/ls.js
@@ -1,10 +1,12 @@
 const { resolve, relative, sep } = require('path')
-const relativePrefix = `.${sep}`
-
 const archy = require('archy')
 const { breadth } = require('treeverse')
 const npa = require('npm-package-arg')
 const { output } = require('proc-log')
+const ArboristWorkspaceCmd = require('../arborist-cmd.js')
+const localeCompare = require('@isaacs/string-locale-compare')('en')
+
+const relativePrefix = `.${sep}`
 
 const _depth = Symbol('depth')
 const _dedupe = Symbol('dedupe')
@@ -17,8 +19,6 @@ const _parent = Symbol('parent')
 const _problems = Symbol('problems')
 const _required = Symbol('required')
 const _type = Symbol('type')
-const ArboristWorkspaceCmd = require('../arborist-cmd.js')
-const localeCompare = require('@isaacs/string-locale-compare')('en')
 
 class LS extends ArboristWorkspaceCmd {
   static description = 'List installed packages'
@@ -219,6 +219,7 @@ class LS extends ArboristWorkspaceCmd {
     return tree
   }
 }
+
 module.exports = LS
 
 const isGitNode = (node) => {
