+++ lib/commands/org.js
@@ -149,4 +149,5 @@ class Org extends BaseCommand {
     }
   }
 }
+
 module.exports = Org
