+++ lib/commands/outdated.js
@@ -6,7 +6,6 @@ const npa = require('npm-package-arg')
 const pickManifest = require('npm-pick-manifest')
 const { output } = require('proc-log')
 const localeCompare = require('@isaacs/string-locale-compare')('en')
-
 const ArboristWorkspaceCmd = require('../arborist-cmd.js')
 
 class Outdated extends ArboristWorkspaceCmd {
@@ -369,4 +368,5 @@ class Outdated extends ArboristWorkspaceCmd {
     return JSON.stringify(out, null, 2)
   }
 }
+
 module.exports = Outdated
