+++ lib/commands/pack.js
@@ -83,4 +83,5 @@ class Pack extends BaseCommand {
     return this.exec([...this.workspacePaths, ...args.filter(a => a !== '.')])
   }
 }
+
 module.exports = Pack
