+++ lib/commands/ping.js
@@ -26,4 +26,5 @@ class Ping extends BaseCommand {
     }
   }
 }
+
 module.exports = Ping
