+++ lib/commands/prefix.js
@@ -11,4 +11,5 @@ class Prefix extends BaseCommand {
     return output.standard(this.npm.prefix)
   }
 }
+
 module.exports = Prefix
