+++ lib/commands/profile.js
@@ -1,11 +1,11 @@
-const inspect = require('util').inspect
+const { inspect } = require('util')
 const { URL } = require('url')
 const { log, output } = require('proc-log')
 const npmProfile = require('npm-profile')
 const qrcodeTerminal = require('qrcode-terminal')
-
 const otplease = require('../utils/otplease.js')
 const readUserInfo = require('../utils/read-user-info.js')
+const BaseCommand = require('../base-command.js')
 
 const qrcode = url =>
   new Promise((resolve) => qrcodeTerminal.generate(url, resolve))
@@ -33,7 +33,6 @@ const writableProfileKeys = [
   'github',
 ]
 
-const BaseCommand = require('../base-command.js')
 class Profile extends BaseCommand {
   static description = 'Change settings on your registry profile'
   static name = 'profile'
@@ -387,4 +386,5 @@ class Profile extends BaseCommand {
     }
   }
 }
+
 module.exports = Profile
