+++ lib/commands/publish.js
@@ -7,18 +7,16 @@ const pacote = require('pacote')
 const npa = require('npm-package-arg')
 const npmFetch = require('npm-registry-fetch')
 const { redactLog: replaceInfo } = require('@npmcli/redact')
-
 const otplease = require('../utils/otplease.js')
 const { getContents, logTar } = require('../utils/tar.js')
-
 // for historical reasons, publishConfig in package.json can contain ANY config
 // keys that npm supports in .npmrc files and elsewhere.  We *may* want to
 // revisit this at some point, and have a minimal set that's a SemVer-major
 // change that ought to get a RFC written on it.
 const { flatten } = require('@npmcli/config/lib/definitions')
 const pkgJson = require('@npmcli/package-json')
-
 const BaseCommand = require('../base-command.js')
+
 class Publish extends BaseCommand {
   static description = 'Publish a package'
   static name = 'publish'
@@ -226,4 +224,5 @@ class Publish extends BaseCommand {
     return manifest
   }
 }
+
 module.exports = Publish
