+++ lib/commands/query.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const { resolve } = require('node:path')
 const BaseCommand = require('../base-command.js')
 const { log, output } = require('proc-log')
