+++ lib/commands/rebuild.js
@@ -2,8 +2,8 @@ const { resolve } = require('path')
 const { output } = require('proc-log')
 const npa = require('npm-package-arg')
 const semver = require('semver')
-
 const ArboristWorkspaceCmd = require('../arborist-cmd.js')
+
 class Rebuild extends ArboristWorkspaceCmd {
   static description = 'Rebuild a package'
   static name = 'rebuild'
@@ -80,4 +80,5 @@ class Rebuild extends ArboristWorkspaceCmd {
     })
   }
 }
+
 module.exports = Rebuild
