+++ lib/commands/root.js
@@ -1,5 +1,6 @@
 const { output } = require('proc-log')
 const BaseCommand = require('../base-command.js')
+
 class Root extends BaseCommand {
   static description = 'Display npm root'
   static name = 'root'
@@ -9,4 +10,5 @@ class Root extends BaseCommand {
     output.standard(this.npm.dir)
   }
 }
+
 module.exports = Root
