+++ lib/commands/run-script.js
@@ -1,7 +1,7 @@
 const { log, output } = require('proc-log')
 const pkgJson = require('@npmcli/package-json')
-
 const BaseCommand = require('../base-command.js')
+
 class RunScript extends BaseCommand {
   static description = 'Run arbitrary package scripts'
   static params = [
