+++ lib/commands/sbom.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const localeCompare = require('@isaacs/string-locale-compare')('en')
 const BaseCommand = require('../base-command.js')
 const { log, output } = require('proc-log')
