+++ lib/commands/search.js
@@ -1,7 +1,6 @@
 const Pipeline = require('minipass-pipeline')
 const libSearch = require('libnpmsearch')
 const { log, output } = require('proc-log')
-
 const formatSearchStream = require('../utils/format-search-stream.js')
 
 const BaseCommand = require('../base-command.js')
@@ -68,4 +67,5 @@ class Search extends BaseCommand {
     log.silly('search', 'search completed')
   }
 }
+
 module.exports = Search
