+++ lib/commands/set.js
@@ -22,4 +22,5 @@ class Set extends BaseCommand {
     return this.npm.exec('config', ['set'].concat(args))
   }
 }
+
 module.exports = Set
