+++ lib/commands/shrinkwrap.js
@@ -69,4 +69,5 @@ class Shrinkwrap extends BaseCommand {
     }
   }
 }
+
 module.exports = Shrinkwrap
