+++ lib/commands/star.js
@@ -2,8 +2,8 @@ const fetch = require('npm-registry-fetch')
 const npa = require('npm-package-arg')
 const { log, output } = require('proc-log')
 const getIdentity = require('../utils/get-identity')
-
 const BaseCommand = require('../base-command.js')
+
 class Star extends BaseCommand {
   static description = 'Mark your favorite packages'
   static name = 'star'
@@ -68,4 +68,5 @@ class Star extends BaseCommand {
     }
   }
 }
+
 module.exports = Star
