+++ lib/commands/stars.js
@@ -1,8 +1,8 @@
 const fetch = require('npm-registry-fetch')
 const { log, output } = require('proc-log')
 const getIdentity = require('../utils/get-identity.js')
-
 const BaseCommand = require('../base-command.js')
+
 class Stars extends BaseCommand {
   static description = 'View packages marked as favorites'
   static name = 'stars'
@@ -35,4 +35,5 @@ class Stars extends BaseCommand {
     }
   }
 }
+
 module.exports = Stars
