+++ lib/commands/team.js
@@ -1,7 +1,6 @@
 const columns = require('cli-columns')
 const libteam = require('libnpmteam')
 const { output } = require('proc-log')
-
 const otplease = require('../utils/otplease.js')
 
 const BaseCommand = require('../base-command.js')
@@ -152,4 +151,5 @@ class Team extends BaseCommand {
     }
   }
 }
+
 module.exports = Team
