+++ lib/commands/token.js
@@ -1,10 +1,9 @@
 const { log, output } = require('proc-log')
 const profile = require('npm-profile')
-
 const otplease = require('../utils/otplease.js')
 const readUserInfo = require('../utils/read-user-info.js')
-
 const BaseCommand = require('../base-command.js')
+
 class Token extends BaseCommand {
   static description = 'Manage your authentication tokens'
   static name = 'token'
@@ -194,4 +193,5 @@ class Token extends BaseCommand {
     return list
   }
 }
+
 module.exports = Token
