+++ lib/commands/unpublish.js
@@ -7,12 +7,12 @@ const pkgJson = require('@npmcli/package-json')
 const { flatten } = require('@npmcli/config/lib/definitions')
 const getIdentity = require('../utils/get-identity.js')
 const otplease = require('../utils/otplease.js')
+const BaseCommand = require('../base-command.js')
 
 const LAST_REMAINING_VERSION_ERROR = 'Refusing to delete the last version of the package. ' +
 'It will block from republishing a new version for 24 hours.\n' +
 'Run with --force to do this.'
 
-const BaseCommand = require('../base-command.js')
 class Unpublish extends BaseCommand {
   static description = 'Remove a package from the registry'
   static name = 'unpublish'
@@ -172,4 +172,5 @@ class Unpublish extends BaseCommand {
     }
   }
 }
+
 module.exports = Unpublish
