+++ lib/commands/unstar.js
@@ -4,4 +4,5 @@ class Unstar extends Star {
   static description = 'Remove an item from your favorite packages'
   static name = 'unstar'
 }
+
 module.exports = Unstar
