+++ lib/commands/update.js
@@ -1,10 +1,8 @@
 const path = require('path')
-
 const { log } = require('proc-log')
-
 const reifyFinish = require('../utils/reify-finish.js')
-
 const ArboristWorkspaceCmd = require('../arborist-cmd.js')
+
 class Update extends ArboristWorkspaceCmd {
   static description = 'Update packages'
   static name = 'update'
@@ -66,4 +64,5 @@ class Update extends ArboristWorkspaceCmd {
     await reifyFinish(this.npm, arb)
   }
 }
+
 module.exports = Update
