+++ lib/commands/version.js
@@ -1,7 +1,6 @@
 const { resolve } = require('node:path')
 const { readFile } = require('node:fs/promises')
 const { output } = require('proc-log')
-
 const BaseCommand = require('../base-command.js')
 
 class Version extends BaseCommand {
