+++ lib/commands/view.js
@@ -9,11 +9,11 @@ const relativeDate = require('tiny-relative-date')
 const semver = require('semver')
 const { inspect } = require('util')
 const { packument } = require('pacote')
+const Queryable = require('../utils/queryable.js')
+const BaseCommand = require('../base-command.js')
 
 const readJson = async file => jsonParse(await readFile(file, 'utf8'))
 
-const Queryable = require('../utils/queryable.js')
-const BaseCommand = require('../base-command.js')
 class View extends BaseCommand {
   static description = 'View registry info'
   static name = 'view'
@@ -406,6 +406,7 @@ class View extends BaseCommand {
     }
   }
 }
+
 module.exports = View
 
 function cleanBlanks (obj) {
