+++ lib/commands/whoami.js
@@ -1,7 +1,7 @@
 const { output } = require('proc-log')
 const getIdentity = require('../utils/get-identity.js')
-
 const BaseCommand = require('../base-command.js')
+
 class Whoami extends BaseCommand {
   static description = 'Display npm username'
   static name = 'whoami'
@@ -14,4 +14,5 @@ class Whoami extends BaseCommand {
     )
   }
 }
+
 module.exports = Whoami
