+++ lib/lifecycle-cmd.js
@@ -16,4 +16,5 @@ class LifecycleCmd extends BaseCommand {
     return this.npm.exec('run-script', [this.constructor.name, ...args])
   }
 }
+
 module.exports = LifecycleCmd
