+++ lib/npm.js
@@ -445,4 +445,5 @@ class Npm {
     this.#display.flushOutput(jsonError)
   }
 }
+
 module.exports = Npm
