+++ lib/package-url-cmd.js
@@ -62,4 +62,5 @@ class PackageUrlCommand extends BaseCommand {
     return (rurl && hostedGitInfo.fromUrl(rurl.replace(/^git\+/, ''))) || null
   }
 }
+
 module.exports = PackageUrlCommand
