+++ lib/utils/did-you-mean.js
@@ -37,4 +37,5 @@ const didYouMean = async (path, scmd) => {
       : `\n\nDid you mean one of these?\n${best.slice(0, 3).join('\n')}`
   return suggestion
 }
+
 module.exports = didYouMean
